ATI (fglrx) PBuffer/X11Display bug workaround/cleanup
    - See https://bugzilla.mozilla.org/show_bug.cgi?id=486277

    - Description:
        - To use PBuffer, a context must be current

        - X11Display cannot be switched while using the PBuffer
          [within one thread]. Hence we shall try harder to reuse
          _the_ user configured X11Display - whenever possible.
          This is actually a good thing, ie cleanup up our
          code again.

    - Changes to workaround/cleanup:
        - GLDrawableFactory* methods 'canCreate*()'
          are changed to 'canCreate*(AbstractGraphicsDevice)'
          to allow pipelining the X11Display.
          This reduces the overhead of using a local TLS X11Display.

    - WindowsDummyWGLDrawable cstr gets the GLProfile as a parameter now,
      this is done while adding X11DummyGLXDrawable - forseeing the
      usecase to query available GLProfiles at startup.

    - X11DummyGLXDrawable added, following the WindowsDummyWGLDrawable path
      to have a dummy GLContext current to fix the ATI bug.

NativeWindow X11:
    - Add XIOErrorHandler to identify the fatal failure
      of closing a Display (-> ATI bug).

Build:
    - Adding ant.jar and ant-junit.jar to the junit compile/run classpath
    -

Misc:
    - Fix: CreateDummyWindow(..) returns a HWND, not a HDC
    - mapToRealGLFunctionName: Added mapping for X11/GLX.
    - X11GLXGraphicsConfigurationFactory: Uncommented dead code 'createDefaultGraphicsConfigurationFBConfig'

Tests: Passed (Linux64bit: NVidia/ATI)

Todo: More tests on ATI, especially multithreading/X11Display usage.

+++ src/jogl/classes/com/jogamp/opengl/impl/windows/wgl/WindowsWGLDrawableFactory.java
@@ -85,12 +85,12 @@ public class WindowsWGLDrawableFactory extends GLDrawableFactoryImpl implements 
 
   private boolean pbufferSupportInitialized = false;
   private boolean canCreateGLPbuffer = false;
-  public boolean canCreateGLPbuffer() {
+  public boolean canCreateGLPbuffer(AbstractGraphicsDevice device) {
     if (!pbufferSupportInitialized) {
       final GLDrawableFactory factory = this;
       Runnable r = new Runnable() {
           public void run() {
-            WindowsDummyWGLDrawable dummyDrawable = new WindowsDummyWGLDrawable(factory);
+            WindowsDummyWGLDrawable dummyDrawable = new WindowsDummyWGLDrawable(factory, null);
             GLContext dummyContext  = dummyDrawable.createContext(null);
             if (dummyContext != null) {
               GLContext lastContext = GLContext.getCurrent();
@@ -123,7 +123,7 @@ public class WindowsWGLDrawableFactory extends GLDrawableFactoryImpl implements 
     final GLDrawableFactory factory = this;
     Runnable r = new Runnable() {
         public void run() {
-          WindowsDummyWGLDrawable dummyDrawable = new WindowsDummyWGLDrawable(factory);
+          WindowsDummyWGLDrawable dummyDrawable = new WindowsDummyWGLDrawable(factory, null);
           WindowsWGLContext       dummyContext  = (WindowsWGLContext) dummyDrawable.createContext(null);
           GLContext lastContext = GLContext.getCurrent();
           if (lastContext != null) {
@@ -136,10 +136,14 @@ public class WindowsWGLDrawableFactory extends GLDrawableFactoryImpl implements 
                                                                            dummyDrawable,
                                                                            dummyWGLExt);
             returnList.add(pbufferDrawable);
-            dummyContext.release();
-            dummyContext.destroy();
-            dummyDrawable.destroy();
           } finally {
+            if(null!=dummyContext) {
+                dummyContext.release();
+                dummyContext.destroy();
+            }
+            if(null!=dummyDrawable) {
+                dummyDrawable.destroy();
+            }
             if (lastContext != null) {
               lastContext.makeCurrent();
             }
@@ -162,7 +166,7 @@ public class WindowsWGLDrawableFactory extends GLDrawableFactoryImpl implements 
     return WindowsExternalWGLContext.create(this, null);
   }
 
-  public boolean canCreateExternalGLDrawable() {
+  public boolean canCreateExternalGLDrawable(AbstractGraphicsDevice device) {
     return true;
   }
 
@@ -222,7 +226,7 @@ public class WindowsWGLDrawableFactory extends GLDrawableFactoryImpl implements 
     return detail;
   }
 
-  public boolean canCreateContextOnJava2DSurface() {
+  public boolean canCreateContextOnJava2DSurface(AbstractGraphicsDevice device) {
     return false;
   }
